doctor-appointment-system/src/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
google-api-python-client
google-auth
tzdata
email-validator
passlib[bcrypt<4.0.0]
python-multipart
python-jose
sqlite-web